Add "keep input" temporary functionality
Allows, along with functionality provided by commit
af7899d00cea783e85129f161d34f6b2a847bd8d, to store the current input
board into the URL. You can copy the URL to obtain a link to the board
which is not automatically processed. This creates suspense for the user
when pasting the URL, as he has to click the "Solve" button himself to
get the answer.

This button is just a testing feature, and should ultimately be
incorporated into the website in a different way.