Change faulty-input detection
Mark all faulty cells for the first faulty cell found in either the
entire row, column or 3x3 grid, depending which on where another faulty
cell is found first. Avoid overwhelming the user with information,
but still be somewhat more useful and complete that the previous method
of only highlighting the first two faulty cells found.

Note that a few unnecessary are removed in `isSafe`.